moved Response Test classes to the same package as other tests
